refactor: reorganize serviceWorker files
- Move serviceWorker.ts to serviceWorker/index.ts
- Move utils/installServiceWorker.ts to serviceWorker/install.ts
- Move utils/message.ts to src/message.ts
- Update all import paths accordingly
